Tarea Semana 6 Java
Código asignado por Scrum Master
