+++ .gitignore
@@ -8,3 +8,4 @@
 /PROGRAMACION/Lección03/AutoboxingUnboxing/nbproject/private/
 /PROGRAMACION/Lección03/AutoboxingUnboxing/build/
 
+/PROGRAMACION/Leccion07/JavaBeans/build/
