+++ PROGRAMACION/Leccion07/JavaBeans/nbproject/genfiles.properties
@@ -4,5 +4,5 @@ build.xml.stylesheet.CRC32=f85dc8f2@1.96.0.48
 # This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
 # Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
 nbproject/build-impl.xml.data.CRC32=edcd0d3c
-nbproject/build-impl.xml.script.CRC32=a732abcb
-nbproject/build-impl.xml.stylesheet.CRC32=f89f7d21@1.96.0.48
+nbproject/build-impl.xml.script.CRC32=b64fff1d
+nbproject/build-impl.xml.stylesheet.CRC32=12e0a6c2@1.105.0.48
