+++ PROGRAMACION/Leccion07/JavaBeans/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\dania\\AppData\\Roaming\\NetBeans\\12.1\\build.properties
+user.properties.file=C:\\Users\\Usuario\\AppData\\Roaming\\NetBeans\\16\\build.properties
