+++ PROGRAMACION/Leccion07/JavaBeans/nbproject/private/private.xml
@@ -3,8 +3,7 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/dania/Tecnicatura3Java/Leccion07/JavaBeans/src/domain/Persona.java</file>
-            <file>file:/C:/Users/dania/Tecnicatura3Java/Leccion07/JavaBeans/src/test/TestJavaBeans.java</file>
+            <file>file:/D:/Usuario/OneDrive/Escritorio/MARTIN/Code-People-Tercer-Semestre/PROGRAMACION/Leccion07/JavaBeans/src/domain/Persona.java</file>
         </group>
     </open-files>
 </project-private>
