+++ PROGRAMACION/Leccion07/JavaBeans/nbproject/project.properties
@@ -71,7 +71,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=JavaBeans
-main.class=
+main.class=test.TestJavaBeans
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
