+++ PROGRAMACION/Leccion08/ManejoExcepciones1/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Mati\\AppData\\Roaming\\NetBeans\\17\\build.properties
+user.properties.file=C:\\Users\\Usuario\\AppData\\Roaming\\NetBeans\\16\\build.properties
