+++ PROGRAMACION/Leccion08/ManejoExcepciones1/nbproject/private/private.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
+    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
+        <group>
+            <file>file:/C:/Users/Usuario/MARTIN/Tec3y4Semestre/PROGRAMACION/New%20Folder/ManejoExcepciones1/src/excepciones/OperacionExcepcion.java</file>
+            <file>file:/C:/Users/Usuario/MARTIN/Tec3y4Semestre/PROGRAMACION/New%20Folder/ManejoExcepciones1/src/test/TestExcepciones.java</file>
+            <file>file:/C:/Users/Usuario/MARTIN/Tec3y4Semestre/PROGRAMACION/New%20Folder/ManejoExcepciones1/src/aritmetica/Aritmetica.java</file>
+        </group>
+    </open-files>
+</project-private>
