+++ PROGRAMACION/Leccion08/ManejoExcepciones1/nbproject/project.properties
@@ -44,8 +44,8 @@ javac.modulepath=
 javac.processormodulepath=
 javac.processorpath=\
     ${javac.classpath}
-javac.source=20
-javac.target=20
+javac.source=17
+javac.target=17
 javac.test.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
