+++ PROGRAMACION/Leccion08/ManejoExcepciones1/src/excepciones/OperacionExcepcion.java
@@ -1,0 +1,11 @@
+
+package excepciones;
+
+public class OperacionExcepcion extends RuntimeException{
+//Contructor
+    public OperacionExcepcion(String message) {
+        super(message);
+    }
+    
+    
+}
