--- PROGRAMACION/Leccion08/ManejodeExcepciones/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\silva\\AppData\\Roaming\\NetBeans\\17\\build.properties
