--- PROGRAMACION/Leccion08/ManejodeExcepciones/nbproject/private/private.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/silva/Documents/GitHub/TecnicaturaGit/Java/Leccion8/ManejodeExcepciones/src/Test/TestExcepciones.java</file>
-        </group>
-    </open-files>
-</project-private>
